Create modelingCCCustomerBehavior.py
First, the data is imported. Categorical features were one hot encoded for both training and test data.
The Not-a-Number values were filled with the median of that feature.
I decided to use Random Forest Classifier because the features of the data were diverse.
Then a Random Forest Classifier is trained using grid search with cross validation. Bootstrapping and stratification are also performed on the data.
Number of Trees and Maximum Depth parameters of the Random Forest was detected using gridsearchCV() function, respectively, in the range 100-140 and 7-12, for all 6 data sets.